Move CompletionList computation into separate file
